notebooks/claimify/pyproject.toml
[project]
name = "claimify-demo"
version = "0.0.0"
description = "Claimify extraction + evaluation demo (notebook workspace)"
requires-python = ">=3.12"
dependencies = [
    "pydantic-handlebars>=0.1.0",
]

[build-system]
requires = ["uv_build>=0.8,<0.9"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "_claimify"
module-root = "."
